Remove redundant calls to determinse data type of properties
