+++ v0/util.js
@@ -3,6 +3,7 @@ const fs = require("fs");
 const path = require("path");
 const _ = require("lodash");
 const set = require("set-value");
+const get = require("get-value");
 const moment = require("moment");
 
 const getMappingConfig = (config, dir) => {
@@ -66,79 +67,6 @@ const toSnakeCase = str => {
     .toLowerCase();
 };
 
-const toSafeDBString = str => {
-  if (parseInt(str[0]) > 0) str = `_${str}`;
-  str = str.replace(/[^a-zA-Z0-9_]+/g, "");
-  return str.substr(0, 127);
-};
-
-// https://www.myintervals.com/blog/2009/05/20/iso-8601-date-validation-that-doesnt-suck/
-// make sure to disable prettier for regex expression
-// prettier-ignore
-const timestampRegex = new RegExp(
-  // eslint-disable-next-line no-useless-escape
-  // /^([\+-]?\d{4})((-)((0[1-9]|1[0-2])(-([12]\d|0[1-9]|3[01])))([T\s]((([01]\d|2[0-3])((:?)[0-5]\d)?|24\:?00)([\.]\d+)?)?(:[0-5]\d([\.]\d+)?)?([zZ]|([\+-])([01]\d|2[0-3]):?([0-5]\d)?)?)?)$/
-  // eslint-disable-next-line no-useless-escape
-  /^([\+-]?\d{4})((-)((0[1-9]|1[0-2])(-([12]\d|0[1-9]|3[01])))([T\s]((([01]\d|2[0-3])((:)[0-5]\d))([\:]\d+)?)?(:[0-5]\d([\.]\d+)?)?([zZ]|([\+-])([01]\d|2[0-3]):?([0-5]\d)?)?)?)$/
-);
-
-function validTimestamp(input) {
-  return timestampRegex.test(input);
-}
-
-// // older implementation with fallback to new Date()
-// function validTimestamp(input) {
-//   // eslint-disable-next-line no-restricted-globals
-//   if (!isNaN(input)) return false;
-//   return new Date(input).getTime() > 0;
-// }
-
-// const timestampFuncFormats = [moment.ISO_8601, moment.RFC_2822];
-
-// const timestampStringFormats = [
-//   "YYYY-MM-DD HH:mm:ss",
-//   "YYYY-MM-DD HH:mm:ss Z",
-//   "YYYY-MM-DD HH:mm:ss.SSS Z",
-//   "YYYY-MM-DDTHH:mm:ss",
-//   "YYYY-MM-DDTHH:mm:ssZ",
-//   "YYYY-MM-DDTHH:mm:ss.SSZ",
-//   "YYYY-MM-DDTHH:mm:ss.SSSZ",
-//   "YYYY-MM-DD",
-//   "MM-DD-YYYY",
-//   "YYYY-MM-DD HH:mm",
-//   "YYYY-MM-DD HH:mm Z",
-//   // ISO_8061 v2.20.0 HTML5 formats
-//   moment.HTML5_FMT.DATETIME_LOCAL,
-//   moment.HTML5_FMT.DATETIME_LOCAL_SECONDS,
-//   moment.HTML5_FMT.DATETIME_LOCAL_MS,
-//   moment.HTML5_FMT.DATE,
-//   moment.HTML5_FMT.TIME,
-//   moment.HTML5_FMT.TIME_SECONDS,
-//   moment.HTML5_FMT.TIME_MS,
-//   moment.HTML5_FMT.WEEK,
-//   moment.HTML5_FMT.MONTH
-// ];
-
-// function validTimestamp(input) {
-//   try {
-//     if (moment(input, timestampStringFormats, true).isValid()) {
-//       return true;
-//     }
-//     for (let index = 0; index < timestampFuncFormats.length; index++) {
-//       var x;
-//       try {
-//         x = moment(input, timestampFuncFormats[index], true).isValid();
-//       } catch (error) {
-//         // ignoring error here;
-//       }
-//       if (x) return true;
-//     }
-//   } catch (error) {
-//     return false;
-//   }
-//   return false;
-// }
-
 const isObject = value => {
   var type = typeof value;
   return value != null && (type == "object" || type == "function");
@@ -180,6 +108,195 @@ const defaultRequestConfig = () => {
   };
 };
 
+// Start of warehouse specific utils
+
+const toSafeDBString = str => {
+  if (parseInt(str[0]) > 0) str = `_${str}`;
+  str = str.replace(/[^a-zA-Z0-9_]+/g, "");
+  return str.substr(0, 127);
+};
+
+// https://www.myintervals.com/blog/2009/05/20/iso-8601-date-validation-that-doesnt-suck/
+// make sure to disable prettier for regex expression
+// prettier-ignore
+const timestampRegex = new RegExp(
+  // eslint-disable-next-line no-useless-escape
+  /^([\+-]?\d{4})((-)((0[1-9]|1[0-2])(-([12]\d|0[1-9]|3[01])))([T\s]((([01]\d|2[0-3])((:)[0-5]\d))([\:]\d+)?)?(:[0-5]\d([\.]\d+)?)?([zZ]|([\+-])([01]\d|2[0-3]):?([0-5]\d)?)?)?)$/
+);
+
+function validTimestamp(input) {
+  return timestampRegex.test(input);
+}
+
+// // older implementation with fallback to new Date()
+// function validTimestamp(input) {
+//   // eslint-disable-next-line no-restricted-globals
+//   if (!isNaN(input)) return false;
+//   return new Date(input).getTime() > 0;
+// }
+
+const getDataType = val => {
+  const type = typeof val;
+  switch (type) {
+    case "number":
+      return Number.isInteger(val) ? "int" : "float";
+    case "boolean":
+      return "boolean";
+    default:
+      break;
+  }
+  if (validTimestamp(val)) {
+    return "datetime";
+  }
+  return "string";
+};
+
+function setFromConfig(input, configJson, columnTypes) {
+  const output = {};
+  Object.keys(configJson).forEach(key => {
+    let val = get(input, key);
+    if (val !== undefined) {
+      datatype = getDataType(val);
+      if (datatype === "datetime") {
+        val = moment(val).format("YYYY-MM-DD hh:mm:ss.SSS"); // supported format in bigquery
+      }
+      output[configJson[key]] = val;
+      columnTypes[configJson[key]] = datatype;
+    }
+  });
+  return output;
+}
+
+function setFromProperties(input, columnTypes, prefix = "") {
+  let output = {};
+  if (!input) return output;
+  Object.keys(input).forEach(key => {
+    if (isObject(input[key])) {
+      output = {
+        ...output,
+        ...setFromProperties(input[key], columnTypes, `${prefix + key}_`)
+      };
+    } else {
+      let val = input[key];
+      datatype = getDataType(val);
+      if (datatype === "datetime") {
+        val = moment(val).format("YYYY-MM-DD hh:mm:ss.SSS"); // supported format in bigquery
+      }
+      output[toSafeDBString(prefix + key)] = val;
+      columnTypes[toSafeDBString(prefix + key)] = datatype;
+    }
+  });
+  return output;
+}
+
+function getColumns(obj, columnTypes) {
+  const columns = { uuid_ts: "datetime" };
+  Object.keys(obj).forEach(key => {
+    columns[toSafeDBString(key)] =
+      columnTypes[obj[key]] || getDataType(obj[key]);
+  });
+  return columns;
+}
+
+function processWarehouseMessage(
+  message,
+  whDefaultConfigJson,
+  whTrackConfigJson
+) {
+  const responses = [];
+  const eventType = message.type.toLowerCase();
+  // store columnTypes as each column is set, so as not to call getDataType again
+  const columnTypes = {};
+  switch (eventType) {
+    case "track": {
+      let result = {
+        ...setFromConfig(message, whDefaultConfigJson, columnTypes),
+        ...setFromProperties(message.context, columnTypes, "context_")
+      };
+      result = {
+        ...result,
+        ...setFromConfig(message, whTrackConfigJson, columnTypes)
+      };
+
+      // set event name
+      result.event = toSnakeCase(result.event_text);
+      columnTypes.event = "string";
+
+      const trackEvent = {
+        ...result,
+        ...setFromProperties(message.properties, columnTypes, ""),
+        ...setFromProperties(message.userProperties, columnTypes, "")
+      };
+
+      const tracksMetadata = {
+        table: "tracks",
+        columns: getColumns(result, columnTypes)
+      };
+      responses.push({ metadata: tracksMetadata, data: result });
+
+      const trackEventMetadata = {
+        table: toSafeDBString(trackEvent.event),
+        columns: getColumns(trackEvent, columnTypes)
+      };
+      responses.push({ metadata: trackEventMetadata, data: trackEvent });
+      break;
+    }
+    case "identify": {
+      const event = {
+        ...setFromProperties(message.context.traits, columnTypes, ""),
+        ...setFromProperties(message.context, columnTypes, "context_")
+      };
+      const usersEvent = { ...event };
+      const identifiesEvent = { ...event };
+
+      usersEvent.id = message.userId;
+      usersEvent.user_id = message.userId;
+      identifiesEvent.user_id = message.userId;
+      identifiesEvent.anonymous_id = message.anonymousId;
+      identifiesEvent.id = message.messageId;
+      // set columntypes
+      columnTypes.id = "string";
+      columnTypes.user_id = "string";
+      columnTypes.anonymous_id = "string";
+
+      const identifiesMetadata = {
+        table: "identifies",
+        columns: getColumns(identifiesEvent, columnTypes)
+      };
+      responses.push({ metadata: identifiesMetadata, data: identifiesEvent });
+
+      const usersMetadata = {
+        table: "users",
+        columns: getColumns(usersEvent, columnTypes)
+      };
+      responses.push({ metadata: usersMetadata, data: usersEvent });
+      break;
+    }
+    case "page":
+    case "screen": {
+      const defaultEvent = {
+        ...setFromConfig(message, whDefaultConfigJson, columnTypes),
+        ...setFromProperties(message.context, columnTypes, "context_")
+      };
+      const event = {
+        ...defaultEvent,
+        ...setFromProperties(message.properties, columnTypes, columnTypes)
+      };
+      const metadata = {
+        table: `${eventType}s`,
+        columns: getColumns(event, columnTypes)
+      };
+      responses.push({ metadata, data: event });
+      break;
+    }
+    default:
+      throw new Error("Unknown event type", eventType);
+  }
+  return responses;
+}
+
+// End of warehouse specific utils
+
 module.exports = {
   getMappingConfig,
   toStringValues,
@@ -191,6 +308,8 @@ module.exports = {
   toSnakeCase,
   toSafeDBString,
   validTimestamp,
+  getDataType,
+  processWarehouseMessage,
   defaultGetRequestConfig,
   defaultPostRequestConfig,
   defaultDeleteRequestConfig,
